move helpers to lib/utils/helpers.js
