--- src/lib/utils/get-object-with-selected-keys.js
@@ -1,11 +1,0 @@
-export default function getObjectWithSelectedKeys(obj, keys) {
-  const newObj = {};
-
-  Object.keys(obj).forEach(k => {
-    if (keys.includes(k)) {
-      newObj[k] = obj[k];
-    }
-  });
-
-  return newObj;
-}
