+++ src/lib/utils/index.js
@@ -1,6 +1,3 @@
 export { default as Seeder } from './seeder';
 export { default as runSeeders } from './run-seeders';
-export {
-  default as getObjectWithSelectedKeys,
-} from './get-object-with-selected-keys';
-export { default as mustContainUserConfig } from './must-contain-user-config';
+export * from './helpers';
